+++ src/Game/Enemy/SlimeEnemy.h
@@ -6,9 +6,9 @@ class GameManager;
 
 class SlimeEnemy : public Enemy {
 private:
-	static constexpr size_t BASE_HEALTH = 100;
-	static constexpr size_t BASE_SPEED = 75;
-	static constexpr size_t BASE_CASH_DROP = 25;
+	static constexpr unsigned int BASE_HEALTH = 100;
+	static constexpr float BASE_SPEED = 75;
+	static constexpr unsigned int BASE_CASH_DROP = 25;
 public:
 	SlimeEnemy();
