+++ src/Game/Enemy/SniperSlimeEnemy.h
@@ -3,9 +3,9 @@
 #include "Game/Enemy.h"
 
 class SniperSlimeEnemy : public Enemy {
-	static constexpr size_t BASE_HEALTH = 150;
-	static constexpr size_t BASE_SPEED = 25;
-	static constexpr size_t BASE_CASH_DROP = 75;
+	static constexpr unsigned int BASE_HEALTH = 150;
+	static constexpr float BASE_SPEED = 25;
+	static constexpr unsigned int BASE_CASH_DROP = 75;
 public:
 	SniperSlimeEnemy();
